+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 ### Fixed
 
 - Start ABR at a lower level corresponding to 480p
+- Let the user choose video quality when ABR is disabled
 
 ## [3.1.6] - 2019-10-24